Several changes.
Added a Rigidbody Component to instantiated static objects to reduce cost on performance when moving them using the GUI.
Fixed a bug that was causing the values for the TextField to be parsed before invoking an update which prevented entering incomplete single floating-point numbers.
Added a starting point to disabling and enabling colliders on static objects when being manipulated.
Changed the Highlight method in StaticObject's class to Manipulate to relate to the new code placed inside.